requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.22
pydantic==2.4.2
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
